@@ -1,58 +1,39 @@
 {
   "name": "@edde746/svelte-icons",
-  "version": "0.0.3-development",
-  "description": "Icon components for svelte",
-  "author": "Chris Ryan",
-  "license": "MIT",
+  "version": "0.1.0",
   "scripts": {
-    "travis-deploy-once": "travis-deploy-once",
-    "semantic-release": "semantic-release",
-    "test": "echo 'no tests here'",
+    "build": "cd scripts && tsc build.ts --esModuleInterop && mv build.js build.cjs && cd .. && node scripts/build.cjs && svelte-kit package",
     "submodule": "git submodule update --init --recursive",
-    "submodule-deinit": "git submodule deinit --all",
-    "build": "ts-node scripts/build.ts",
-    "dev": "watch 'yarn build' components icons",
-    "commit": "git-cz"
+    "submodule-deinit": "git submodule deinit --all"
   },
   "devDependencies": {
-    "@types/node": "^12.6.8",
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/kit": "next",
     "@types/cheerio": "^0.22.12",
+    "@types/node": "^12.20.52",
     "@types/rimraf": "^2.0.2",
-    "commitizen": "^3.0.5",
-    "cz-conventional-changelog": "2.1.0",
-    "semantic-release": "^15.9.9",
-    "travis-deploy-once": "^5.0.3",
-    "commander": "^2.20.0",
-    "prettier": "^1.18.2",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3",
     "camelcase": "^5.0.0",
     "cheerio": "^1.0.0-rc.2",
+    "commander": "^2.20.0",
+    "commitizen": "^3.0.5",
     "copy": "^0.3.2",
+    "cz-conventional-changelog": "2.1.0",
     "feather-icons": "^4.22.1",
     "glob-promise": "^3.4.0",
     "ionicons": "^4.6.2",
     "octicons": "^8.5.0",
+    "prettier": "^1.18.2",
     "rimraf": "^2.6.3",
+    "semantic-release": "^15.9.9",
+    "svelte": "^3.44.0",
+    "svelte-check": "^2.2.6",
+    "svelte-preprocess": "^4.10.1",
+    "svelte2tsx": "^0.5.9",
+    "travis-deploy-once": "^5.0.3",
+    "tslib": "^2.3.1",
+    "tslint": "^5.18.0",
+    "typescript": "~4.6.2",
     "watch": "^1.0.2"
   },
-  "files": [
-    "wi",
-    "ti",
-    "md",
-    "io",
-    "go",
-    "gi",
-    "fi",
-    "fa",
-    "di",
-    "hi",
-    "components"
-  ],
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  }
+  "type": "module"
 }
