@@ -1,17 +1,13 @@
 import cheerio from 'cheerio';
 import glob from 'glob-promise';
 import path from 'path';
-import fs = require('fs');
-import { promisify } from 'util';
+import { promises as fs } from 'fs';
 import camelcase from 'camelcase';
 import rimraf from 'rimraf';
 
-const fsdir = promisify(fs.mkdir);
-const writeFile = promisify(fs.writeFile);
-
 const rootDir = path.resolve(__dirname, '..');
 const iconsDir = path.resolve(rootDir, 'icons');
-const DIST = path.resolve(rootDir, '.');
+const DIST = path.resolve(rootDir, '.', 'src', 'lib');
 
 interface Content {
   files: string;
@@ -31,11 +27,11 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(iconsDir, 'fontawesome/svgs/+(brands|solid)/*.svg'),
-        formatter: name => `Fa${name}`,
+        formatter: (name) => `Fa${name}`,
       },
       {
         files: path.resolve(iconsDir, 'fontawesome/svgs/regular/*.svg'),
-        formatter: name => `FaReg${name}`,
+        formatter: (name) => `FaReg${name}`,
       },
     ],
   },
@@ -44,11 +40,8 @@ const icons: Icon[] = [
     name: 'Ionicons',
     contents: [
       {
-        files: path.resolve(
-          rootDir,
-          'node_modules/ionicons/dist/collection/icon/svg/*.svg'
-        ),
-        formatter: name => `Io${name}`,
+        files: path.resolve(rootDir, 'node_modules/ionicons/dist/collection/icon/svg/*.svg'),
+        formatter: (name) => `Io${name}`,
       },
     ],
   },
@@ -57,11 +50,8 @@ const icons: Icon[] = [
     name: 'Material Design icons',
     contents: [
       {
-        files: path.resolve(
-          iconsDir,
-          'material-design-icons/*/svg/production/*_24px.svg'
-        ),
-        formatter: name => name.replace(/Ic(\w+)24px/i, 'Md$1'),
+        files: path.resolve(iconsDir, 'material-design-icons/*/svg/production/*_24px.svg'),
+        formatter: (name) => name.replace(/Ic(\w+)24px/i, 'Md$1'),
       },
     ],
   },
@@ -71,7 +61,7 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(iconsDir, 'typicons/src/svg/*.svg'),
-        formatter: name => `Ti${name}`,
+        formatter: (name) => `Ti${name}`,
       },
     ],
   },
@@ -81,7 +71,7 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(rootDir, 'node_modules/octicons/build/svg/*.svg'),
-        formatter: name => `Go${name}`,
+        formatter: (name) => `Go${name}`,
       },
     ],
   },
@@ -109,60 +99,45 @@ const icons: Icon[] = [
           iconsDir,
           'game-icons-inverted/+(carl-olsen|andymeneely|cathelineau|darkzaitzev|delapouite|faithtoken|generalace135|guard13007|heavenly-dog|irongamer|john-colburn|kier-heyl|lorc|lord-berandas|quoting|rihlsul|sbed|skoll|sparker|spencerdub|zajkonur)/originals/svg/000000/transparent/*.svg'
         ),
-        formatter: name => `Gi${name}`,
+        formatter: (name) => `Gi${name}`,
       },
       {
         files: path.resolve(
           iconsDir,
           'game-icons-inverted/+(zeromancer|willdabeast|)/deviations/svg/000000/transparent/*.svg'
         ),
-        formatter: name => `Gi${name}`,
+        formatter: (name) => `Gi${name}`,
       },
       {
-        files: path.resolve(
-          iconsDir,
-          'game-icons-inverted/+(john-redman)/hands/svg/000000/transparent/*.svg'
-        ),
-        formatter: name => `Gi${name}`,
+        files: path.resolve(iconsDir, 'game-icons-inverted/+(john-redman)/hands/svg/000000/transparent/*.svg'),
+        formatter: (name) => `Gi${name}`,
       },
       {
-        files: path.resolve(
-          iconsDir,
-          'game-icons-inverted/+(lucasms)/equipment/svg/000000/transparent/*.svg'
-        ),
-        formatter: name => `Gi${name}`,
+        files: path.resolve(iconsDir, 'game-icons-inverted/+(lucasms)/equipment/svg/000000/transparent/*.svg'),
+        formatter: (name) => `Gi${name}`,
       },
       {
-        files: path.resolve(
-          iconsDir,
-          'game-icons-inverted/+(priorblue)/batteries/svg/000000/transparent/*.svg'
-        ),
-        formatter: name => `Gi${name}`,
+        files: path.resolve(iconsDir, 'game-icons-inverted/+(priorblue)/batteries/svg/000000/transparent/*.svg'),
+        formatter: (name) => `Gi${name}`,
       },
       {
-        files: path.resolve(
-          iconsDir,
-          'game-icons-inverted/+(viscious-speed)/abstract/svg/000000/transparent/*.svg'
-        ),
-        formatter: name => `Gi${name}`,
+        files: path.resolve(iconsDir, 'game-icons-inverted/+(viscious-speed)/abstract/svg/000000/transparent/*.svg'),
+        formatter: (name) => `Gi${name}`,
       },
       {
         files: path.resolve(
           iconsDir,
           'game-icons-inverted/+(various-artists)/public-domain/svg/000000/transparent/*.svg'
         ),
-        formatter: name => `Gi${name}`,
+        formatter: (name) => `Gi${name}`,
       },
       {
-        files: path.resolve(
-          iconsDir,
-          'game-icons-inverted/+(felbrigg)/arrows/svg/000000/transparent/*.svg'
-        ),
-        formatter: name => `Gi${name}`,
+        files: path.resolve(iconsDir, 'game-icons-inverted/+(felbrigg)/arrows/svg/000000/transparent/*.svg'),
+        formatter: (name) => `Gi${name}`,
       },
       {
         files: path.resolve(iconsDir, 'game-icons-inverted/aussiesim/*.svg'),
-        formatter: name => `Gi${name}`,
+        formatter: (name) => `Gi${name}`,
       },
     ],
   },
@@ -172,7 +147,7 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(iconsDir, 'weather-icons/svg/*.svg'),
-        formatter: name => name,
+        formatter: (name) => name,
       },
     ],
   },
@@ -182,7 +157,7 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(iconsDir, 'devicons/!SVG/*.svg'),
-        formatter: name => `Di${name}`,
+        formatter: (name) => `Di${name}`,
       },
     ],
   },
@@ -192,32 +167,35 @@ const icons: Icon[] = [
     contents: [
       {
         files: path.resolve(iconsDir, 'heroicons/src/solid/*.svg'),
-        formatter: name => `HeroSolid${name}`,
+        formatter: (name) => `Hi${name}Solid`,
       },
       {
         files: path.resolve(iconsDir, 'heroicons/src/outline/*.svg'),
-        formatter: name => `HeroOutline${name}`,
+        formatter: (name) => `Hi${name}Outline`,
       },
     ],
   },
 ];
 
 const mkdir = (dir: string) => {
-  const exists = fs.existsSync(dir);
+  const exists = fs
+    .access(dir)
+    .then(() => true)
+    .catch(() => false);
   if (exists) {
     rimraf.sync(dir);
   }
-  return fsdir(dir);
+  return fs.mkdir(dir);
 };
 
-const write = (filePath: string[], content: string) =>
-  writeFile(path.resolve(DIST, ...filePath), content, 'utf8');
+const write = (filePath: string[], content: string) => fs.writeFile(path.resolve(DIST, ...filePath), content, 'utf8');
 
 const getIconFiles = async (content: Content) => glob(content.files);
 
 interface Element {
   tag: string;
   attr: {
+    fill: any;
     viewBox: string;
     d: string;
   };
@@ -234,12 +212,10 @@ async function convertIconData(svg: string) {
     attribs &&
     Object.keys(attribs)
       .filter(
-        name =>
+        (name) =>
           ![
             'class',
-            ...(tagName === 'svg'
-              ? ['xmlns', 'xmlns:xlink', 'xml:space', 'width', 'height']
-              : []), // if tagName is svg remove size attributes
+            ...(tagName === 'svg' ? ['xmlns', 'xmlns:xlink', 'xml:space', 'width', 'height'] : []), // if tagName is svg remove size attributes
           ].includes(name)
       )
       .reduce((obj, name) => {
@@ -257,16 +233,13 @@ async function convertIconData(svg: string) {
         return obj;
       }, {});
 
-  const elementToTree: any = (element: Cheerio) =>
+  const elementToTree: any = (element: any) =>
     element
-      .filter((_, e) => !!e.tagName && !['style'].includes(e.tagName))
-      .map((_, e) => ({
+      .filter((_: any, e: any) => !!e.tagName && !['style'].includes(e.tagName))
+      .map((_: any, e: any) => ({
         tag: e.tagName,
         attr: attrConverter(e.attribs, e.tagName),
-        child:
-          e.children && e.children.length
-            ? elementToTree(cheerio(e.children))
-            : undefined,
+        child: e.children && e.children.length ? elementToTree(cheerio(e.children)) : undefined,
       }))
       .get();
 
@@ -277,24 +250,11 @@ async function convertIconData(svg: string) {
 async function dirInit() {
   for (const icon of icons) {
     await mkdir(path.resolve(DIST, icon.id));
-    await write([icon.id, 'index.js'], '// THIS FILE IS AUTO GENERATED\n');
-    await write(
-      [icon.id, 'package.json'],
-      JSON.stringify(
-        {
-          sideEffects: false,
-          module: './index.js',
-        },
-        null,
-        2
-      ) + '\n'
-    );
-    await mkdir(path.resolve(icon.id));
+    await write([icon.id, 'index.ts'], '// THIS FILE IS AUTO GENERATED\n');
   }
 }
 
 async function writeIconModule(icon: Icon) {
-  const appendFile = promisify(fs.appendFile);
   const filenames = new Set(); // for remove duplicate
   const filepaths = new Set();
   for (const content of icon.contents) {
@@ -302,20 +262,19 @@ async function writeIconModule(icon: Icon) {
 
     for (let i = 0; i < files.length; i++) {
       const file = files[i];
-      const svgStr = await promisify(fs.readFile)(file, 'utf8');
+      const svgStr = await fs.readFile(file, 'utf8');
       const iconData = await convertIconData(svgStr);
 
       const rawName = path.basename(file, path.extname(file));
 
       const pascalName = camelcase(rawName, { pascalCase: true });
-      const name =
-        (content.formatter && content.formatter(pascalName)) || pascalName;
+      const name = (content.formatter && content.formatter(pascalName)) || pascalName;
       if (filenames.has(name) || filepaths.has(name)) continue;
       filenames.add(name);
 
       // write like: module/fa/index.esm.js
-      await appendFile(
-        path.resolve(DIST, icon.id, 'index.js'),
+      await fs.appendFile(
+        path.resolve(DIST, icon.id, 'index.ts'),
         `export { default as ${name} } from './${name}.svelte';\n`,
         'utf8'
       );
@@ -333,11 +292,11 @@ async function writeIconModule(icon: Icon) {
       await write(
         [icon.id, `${name}.svelte`],
         `<script>
-        import IconBase from '../components/IconBase.svelte';
+        import Icon from '../Icon.svelte';
         </script>
-        <IconBase viewBox="${viewBox}" {...$$props}>
+        <Icon viewBox="${viewBox}" ${attr?.fill ? `fill="${attr.fill}"` : ''} {...$$props}>
           ${svgContent}
-        </IconBase>
+        </Icon>
       `
       );
 
